# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IAEDU API",
        "description": "API para el sistema educativo IAEDU",
        "contact": {
            "email": "admin@iaedu.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/attendance": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Obtener lista de asistencias",
                "operationId": "0ff39c7541f625e763922bd0aaa25ea8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Filtrar por estudiante",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filtrar por fecha",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (presente, ausente, tardanza)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Incluir relaciones (student)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de asistencias obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Attendance records retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Attendance"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Crear nueva asistencia",
                "operationId": "08e675eac4736ac76f8fe56643406d2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "fecha",
                                    "estado"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "presente",
                                            "ausente",
                                            "tardanza"
                                        ],
                                        "example": "presente"
                                    },
                                    "observaciones": {
                                        "type": "string",
                                        "example": "Llegó a tiempo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asistencia creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Attendance record created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Attendance"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
        "/api/attendance/statistics": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Obtener estadísticas de asistencia",
                "operationId": "292f3236077276a88f547ab13d893c1a",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Filtrar por estudiante",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Attendance statistics retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_days": {
                                                    "type": "integer"
                                                },
                                                "present": {
                                                    "type": "integer"
                                                },
                                                "absent": {
                                                    "type": "integer"
                                                },
                                                "late": {
                                                    "type": "integer"
                                                },
                                                "attendance_rate": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Attendance": {
                "title": "Attendance",
                "description": "Modelo de asistencia del sistema",
                "properties": {
                    "id": {
                        "description": "ID único de la asistencia",
                        "type": "integer",
                        "example": 1
                    },
                    "student_id": {
                        "description": "ID del estudiante",
                        "type": "integer",
                        "example": 1
                    },
                    "subject_id": {
                        "description": "ID de la materia",
                        "type": "integer",
                        "example": 1
                    },
                    "date": {
                        "description": "Fecha de la asistencia",
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "status": {
                        "description": "Estado de la asistencia",
                        "type": "string",
                        "enum": [
                            "present",
                            "absent",
                            "late"
                        ],
                        "example": "present"
                    },
                    "justification_type": {
                        "description": "Tipo de justificación",
                        "type": "string",
                        "example": "medical"
                    },
                    "justification_document": {
                        "description": "Documento de justificación",
                        "type": "string",
                        "example": "document.pdf"
                    },
                    "observations": {
                        "description": "Observaciones adicionales",
                        "type": "string",
                        "example": "Llegó a tiempo"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}